--- app/models/concerns/businesses/business_forms/notifications.rb
@@ -1,16 +1,0 @@
-module Businesses
-  module BusinessForms
-    module Notifications
-      def save_and_notify
-        if save
-          send_admin_notification
-          true
-        end
-      end
-
-      def send_admin_notification
-        Admin::NewHireFormNotification.with(form: self).deliver_later(User.admin)
-      end
-    end
-  end
-end
