+++ app/models/forms.rb
@@ -1,0 +1,5 @@
+module Forms
+  def self.table_name_prefix
+    "forms_"
+  end
+end
