+++ db/migrate/20230320214634_create_forms_businesses_hires.rb
@@ -1,6 +1,6 @@
-class CreateBusinessForms < ActiveRecord::Migration[7.0]
+class CreateFormsBusinessesHires < ActiveRecord::Migration[7.0]
   def change
-    create_table :business_forms do |t|
+    create_table :forms_businesses_hires do |t|
       t.belongs_to :business, null: false, foreign_key: true
       t.text :billing_address, null: false
       t.string :developer_name, null: false
