+++ db/schema.rb
@@ -61,20 +61,6 @@ ActiveRecord::Schema[7.0].define(version: 2023_03_19_212906) do
     t.datetime "updated_at", null: false
   end
 
-  create_table "business_forms", force: :cascade do |t|
-    t.bigint "business_id", null: false
-    t.text "billing_address", null: false
-    t.string "developer_name", null: false
-    t.string "position", null: false
-    t.date "start_date", null: false
-    t.integer "annual_salary", null: false
-    t.integer "employment_type", null: false
-    t.text "feedback"
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
-    t.index ["business_id"], name: "index_business_forms_on_business_id"
-  end
-
   create_table "analytics_search_queries", force: :cascade do |t|
     t.string "search_query"
     t.integer "specialty_ids", default: [], null: false, array: true
@@ -147,6 +133,20 @@ ActiveRecord::Schema[7.0].define(version: 2023_03_19_212906) do
     t.index ["public_profile_key"], name: "index_developers_on_public_profile_key", unique: true
     t.index ["textsearchable_index_col"], name: "textsearchable_index", using: :gin
     t.index ["user_id"], name: "index_developers_on_user_id"
+  end
+
+  create_table "forms_businesses_hires", force: :cascade do |t|
+    t.bigint "business_id", null: false
+    t.text "billing_address", null: false
+    t.string "developer_name", null: false
+    t.string "position", null: false
+    t.date "start_date", null: false
+    t.integer "annual_salary", null: false
+    t.integer "employment_type", null: false
+    t.text "feedback"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.index ["business_id"], name: "index_forms_businesses_hires_on_business_id"
   end
 
   create_table "hired_forms", force: :cascade do |t|
